// src/types/types.ts
export type User = {
  id: string;
  instance_id: string | null;
  aud: string | null;
  email: string;
  password: string;
  email_confirmed_at: string | null;
  invited_at: string | null;
  confirmation_token: string | null;
  confirmation_sent_at: string | null;
  recovery_token: string | null;
  recovery_sent_at: string | null;
  email_change_token_new: string | null;
  email_change: string | null;
  email_change_sent_at: string | null;
  last_sign_in_at: string | null;
  metadata: unknown | null;
  flags: unknown | null;
  picture_url: string | null;
  name: string;
  username: string;
  phone: string | null;
  phone_verified: boolean;
  birthdate: string | null;
  gender: string | null;
  slug: string | null;
  provider_id: string | null;
  provider: string | null;
  provider_type: string | null;
  nationality: string | null;
  terms_and_policies: boolean;
  street: string | null;
  neighborhood: string | null;
  city: string | null;
  state_province: string | null;
  country: string | null;
  postal_code: string | null;
  language: string | null;
  currency: string | null;
  website: string | null;
  work_local: string | null;
  occupation: string | null;
  company: string | null;
  academic_level: string | null;
  role?: 'buyer' | 'seller' | 'admin';
  createdAt: string;
  updatedAt: string;
};

export type ChatMessage = {
  id?: string; 
  user_id?: string; 
  receiver_id?: string;
  username?: string; 
  message: string;  
  viewed?: boolean; 
  created_at: string;
};